refactor: less redundancy in format crate names
