+++ wdl/src/bin/wdl.rs
@@ -22,7 +22,7 @@ use wdl::ast::Diagnostic;
 use wdl::ast::Document;
 use wdl::ast::SyntaxNode;
 use wdl::ast::Validator;
-use wdl::format::format::format_document;
+use wdl::format::format_document;
 use wdl::lint::LintVisitor;
 use wdl_analysis::AnalysisEngine;
 use wdl_analysis::AnalysisResult;
